pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unity-docs-mcp"
version = "0.1.0"
description = "Local Unity 6.3 documentation bake/index MCP server with hybrid retrieval."
authors = [{ name = "UnityRAG" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "tqdm>=4.66.4",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "mcp>=1.25.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.4"]

[project.scripts]
unitydocs-setup = "unity_docs_mcp.setup.ensure_artifacts:main"
unitydocs-bake = "unity_docs_mcp.bake.bake_cli:main"
unitydocs-index = "unity_docs_mcp.index.index_cli:main"
unitydocs-mcp = "unity_docs_mcp.mcp_server:main"

[tool.setuptools.packages.find]
where = ["src"]
